+++ src/data/bucket/fill_extrusion_bucket.test.ts
@@ -14,11 +14,11 @@ function createFillExtrusionBucket({id, layout, paint, globalState, availableIma
         type: 'fill-extrusion',
         layout,
         paint
-    } as LayerSpecification);
-    layer.recalculate({zoom: 0, zoomHistory: {} as ZoomHistory, globalState} as EvaluationParameters,
+    } as LayerSpecification, globalState);
+    layer.recalculate({zoom: 0, zoomHistory: {} as ZoomHistory} as EvaluationParameters,
         availableImages as Array<string>);
 
-    return new FillExtrusionBucket({layers: [layer], globalState} as BucketParameters<FillExtrusionStyleLayer>);
+    return new FillExtrusionBucket({layers: [layer]} as BucketParameters<FillExtrusionStyleLayer>);
 }
 
 describe('FillExtrusionBucket', () => {