+++ src/source/worker_source.ts
@@ -39,7 +39,6 @@ export type WorkerTileParameters = TileParameters & {
     collectResourceTiming?: boolean;
     returnDependencies?: boolean;
     subdivisionGranularity: SubdivisionGranularitySetting;
-    globalState: Record<string, any>;
 };
 
 /**