+++ test/unit/lib/create_symbol_layer.ts
@@ -10,15 +10,14 @@ export function createSymbolBucket(layerId, font, text, collisionBoxArray) {
         type: 'symbol',
         layout: {'text-font': [font], 'text-field': text},
         filter: featureFilter(undefined)
-    } as any as LayerSpecification);
+    } as any as LayerSpecification, {});
     layer.recalculate({zoom: 0, zoomHistory: {}} as EvaluationParameters, undefined);
 
     return new SymbolBucket({
         overscaling: 1,
         zoom: 0,
         collisionBoxArray,
-        layers: [layer],
-        globalState: {}
+        layers: [layer]
     } as BucketParameters<SymbolStyleLayer>);
 }
 
@@ -28,7 +27,7 @@ export function createSymbolIconBucket(layerId, iconProperty, collisionBoxArray)
         type: 'symbol',
         layout: {'icon-image': ['get', iconProperty]},
         filter: featureFilter(undefined)
-    } as any as LayerSpecification);
+    } as any as LayerSpecification, {});
     layer.recalculate({zoom: 0, zoomHistory: {}} as EvaluationParameters, undefined);
 
     return new SymbolBucket({